+++ src/shared/components/AddonShell/index.js
@@ -1,0 +1,14 @@
+/**
+ * Add-on Shell Components.
+ *
+ * Shared components for building consistent add-on UIs.
+ *
+ * @package VirtualMediaFolders
+ */
+
+import './addon-shell.scss';
+
+export { AddonShell, default } from './AddonShell';
+export { StatsCard } from './StatsCard';
+export { SubTabNav, SUB_TABS } from './SubTabNav';
+export { StatusIndicator } from './StatusIndicator';